+++ aldryn_newsblog/cms_toolbars.py
@@ -1,8 +1,10 @@
-from django.urls import reverse
+from django.contrib.contenttypes.models import ContentType
+from django.urls import Resolver404, resolve, reverse
 from django.utils.translation import get_language_from_request
 from django.utils.translation import gettext as _
 from django.utils.translation import override
 
+from cms.models.contentmodels import PageContent
 from cms.toolbar.items import ButtonList
 from cms.toolbar_base import CMSToolbar
 from cms.toolbar_pool import toolbar_pool
@@ -77,6 +79,27 @@ class NewsBlogToolbar(CMSToolbar):
                     redirect_url = self.get_on_delete_redirect_url(obj, language=language)
                     url = get_admin_url('aldryn_newsblog_article_delete', [obj.pk])
                     menu.add_modal_item(_('Delete this article'), url=url, on_close=redirect_url)
+        try:
+            self.enable_edit_page_content(language)
+        except (Resolver404, ContentType.DoesNotExist, PageContent.DoesNotExist):
+            pass
+
+    def enable_edit_page_content(self, language: str) -> None:
+        """Enable edit PageContent."""
+        view_func, args, kwargs = resolve(self.request.path)
+        if view_func.__name__ in ("render_object_edit", "render_object_preview") and len(args) == 2:
+            content_type_id, object_id = args
+            content_type = ContentType.objects.get_for_id(content_type_id)
+            model = content_type.model_class()
+            if not issubclass(model, PageContent):
+                return
+            if "render_object_edit":
+                content_type_obj = model.admin_manager.select_related("page").get(pk=object_id)
+            else:
+                content_type_obj = model.objects.select_related("page").get(pk=object_id)
+        else:
+            content_type_obj = PageContent.objects.get(page=self.request.current_page, language=language)
+        self.toolbar.set_object(content_type_obj)
 
     def post_template_populate(self):
         # Disable call self.add_wizard_button().