+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 Unreleased
 ==========
 
+* Enable edit PageContent.
 * Fix template pagination for view search.
 * Fix request.is_ajax in views.